# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EmotionSense API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DetectFaceBody {
  /** Base64-encoded image data */
  image: string;
}

export interface DetectVoiceBody {
  /** Base64-encoded audio data */
  audio: string;
}

export interface EmotionResult {
  /** Detected emotion label */
  emotion: string;
  /** Confidence score between 0 and 1 */
  confidence: number;
  /**
   * ID of the saved log entry
   * @nullable
   */
  logId: number | null;
  /** ISO timestamp of detection */
  timestamp: string;
}

export interface EmotionLog {
  id: number;
  /** @nullable */
  faceEmotion: string | null;
  /** @nullable */
  voiceEmotion: string | null;
  /** @nullable */
  faceConfidence: number | null;
  /** @nullable */
  voiceConfidence: number | null;
  createdAt: string;
}

export type EmotionStatsFaceEmotionCounts = { [key: string]: number };

export type EmotionStatsVoiceEmotionCounts = { [key: string]: number };

export interface EmotionStats {
  totalDetections: number;
  faceEmotionCounts: EmotionStatsFaceEmotionCounts;
  voiceEmotionCounts: EmotionStatsVoiceEmotionCounts;
  /** @nullable */
  mostCommonFaceEmotion: string | null;
  /** @nullable */
  mostCommonVoiceEmotion: string | null;
}

export type GetEmotionLogsParams = {
  limit?: number;
};