# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EmotionSense API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Analyzes a base64 image and returns detected facial emotion
 * @summary Detect face emotion
 */
export const DetectFaceBody = zod.object({
  image: zod.string().describe("Base64-encoded image data"),
});

export const DetectFaceResponse = zod.object({
  emotion: zod.string().describe("Detected emotion label"),
  confidence: zod.number().describe("Confidence score between 0 and 1"),
  logId: zod.number().nullable().describe("ID of the saved log entry"),
  timestamp: zod.string().describe("ISO timestamp of detection"),
});

/**
 * Analyzes base64 audio and returns detected voice emotion
 * @summary Detect voice emotion
 */
export const DetectVoiceBody = zod.object({
  audio: zod.string().describe("Base64-encoded audio data"),
});

export const DetectVoiceResponse = zod.object({
  emotion: zod.string().describe("Detected emotion label"),
  confidence: zod.number().describe("Confidence score between 0 and 1"),
  logId: zod.number().nullable().describe("ID of the saved log entry"),
  timestamp: zod.string().describe("ISO timestamp of detection"),
});

/**
 * Returns the last N emotion detection logs
 * @summary Get emotion history
 */
export const getEmotionLogsQueryLimitDefault = 5;

export const GetEmotionLogsQueryParams = zod.object({
  limit: zod.coerce.number().default(getEmotionLogsQueryLimitDefault),
});

export const GetEmotionLogsResponseItem = zod.object({
  id: zod.number(),
  faceEmotion: zod.string().nullable(),
  voiceEmotion: zod.string().nullable(),
  faceConfidence: zod.number().nullable(),
  voiceConfidence: zod.number().nullable(),
  createdAt: zod.string(),
});
export const GetEmotionLogsResponse = zod.array(GetEmotionLogsResponseItem);

/**
 * Returns aggregated counts of detected emotions
 * @summary Get emotion statistics
 */
export const GetEmotionStatsResponse = zod.object({
  totalDetections: zod.number(),
  faceEmotionCounts: zod.record(zod.string(), zod.number()),
  voiceEmotionCounts: zod.record(zod.string(), zod.number()),
  mostCommonFaceEmotion: zod.string().nullable(),
  mostCommonVoiceEmotion: zod.string().nullable(),
});