# lib/api-zod/src/generated/types/detectFaceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EmotionSense API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DetectFaceBody {
  /** Base64-encoded image data */
  image: string;
}