# lib/api-zod/src/generated/types/detectVoiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EmotionSense API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DetectVoiceBody {
  /** Base64-encoded audio data */
  audio: string;
}