# lib/api-zod/src/generated/types/emotionLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EmotionSense API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmotionLog {
  id: number;
  /** @nullable */
  faceEmotion: string | null;
  /** @nullable */
  voiceEmotion: string | null;
  /** @nullable */
  faceConfidence: number | null;
  /** @nullable */
  voiceConfidence: number | null;
  createdAt: string;
}