# lib/api-zod/src/generated/types/emotionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EmotionSense API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmotionResult {
  /** Detected emotion label */
  emotion: string;
  /** Confidence score between 0 and 1 */
  confidence: number;
  /**
   * ID of the saved log entry
   * @nullable
   */
  logId: number | null;
  /** ISO timestamp of detection */
  timestamp: string;
}