# lib/api-zod/src/generated/types/emotionStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EmotionSense API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmotionStatsFaceEmotionCounts } from "./emotionStatsFaceEmotionCounts";
import type { EmotionStatsVoiceEmotionCounts } from "./emotionStatsVoiceEmotionCounts";

export interface EmotionStats {
  totalDetections: number;
  faceEmotionCounts: EmotionStatsFaceEmotionCounts;
  voiceEmotionCounts: EmotionStatsVoiceEmotionCounts;
  /** @nullable */
  mostCommonFaceEmotion: string | null;
  /** @nullable */
  mostCommonVoiceEmotion: string | null;
}